requirements.txt
requests>=2.31.0
apscheduler>=3.10.4
numpy>=1.26.0
scipy>=1.11.0
pandas>=2.1.0
python-dotenv>=1.0.0
